♨

+++ ubuntu/init.sh
@@ -30,10 +30,15 @@ if [ -n "$GFW" ]; then
   sed -i "s/archive.ubuntu.com/mirrors.163.com/g" /etc/apt/sources.list
   rsync -avI $ZHOS/ /
   source $ZHOS/root/.export
+  if ! [ -x "$(command -v pip3)" ]; then
   apt-get install -y python
+  fi
   pip3 install apt-select
   apt-select --country CN
-  mv /root/sources.list /etc/apt/
+  if [ -f "/root/sources.list" ]; then
+    mv /root/sources.list /etc/apt/
+  fi
+
 fi
 
 systemctl stop snapd || true
